+++ coldfront/core/project/views_/renewal_views/approval_views.py
@@ -11,8 +11,7 @@ from coldfront.core.project.utils_.renewal_utils import AllocationRenewalProcess
 from coldfront.core.project.utils_.renewal_utils import allocation_renewal_request_denial_reason
 from coldfront.core.project.utils_.renewal_utils import allocation_renewal_request_latest_update_timestamp
 from coldfront.core.project.utils_.renewal_utils import allocation_renewal_request_state_status
-from coldfront.core.project.utils_.renewal_utils import get_renewal_survey
-from coldfront.core.project.utils_.renewal_utils import get_gspread_wks
+from coldfront.core.project.utils_.gspread import get_survey_response
 from coldfront.core.resource.utils_.allowance_utils.computing_allowance import ComputingAllowance
 from coldfront.core.resource.utils_.allowance_utils.interface import ComputingAllowanceInterface
 from coldfront.core.utils.common import display_time_zone_current_date
@@ -101,8 +100,10 @@ class AllocationRenewalRequestMixin(object):
             logger.exception(e)
             messages.error(self.request, self.error_message)
             context['allocation_amount'] = 'Failed to compute.'
-        context['survey_response'] = \
-            self.get_google_renewal_survey_responses()
+        context['survey_response'] = get_survey_response(
+            self.request_obj.allocation_period.name,
+            self.request_obj.post_project.name,
+            self.request_obj.pi.username)
         context['has_survey_answers'] = bool(context['survey_response'])
         return context
 
@@ -123,32 +124,6 @@ class AllocationRenewalRequestMixin(object):
                 num_service_units, self.request_obj.request_time,
                 self.allocation_period_obj)
         return num_service_units
-
-    def get_google_renewal_survey_responses(self):
-        """ Takes information from the request object and returns an
-         iterable of tuples representing the requester's survey answers. If no
-         answer is detected, return None. The format of the tuple:
-         ( question: string, answer: string ).  """
-        gform_info = get_renewal_survey(self.request_obj.allocation_period.name)
-        wks = get_gspread_wks(gform_info['sheet_id'], 0)
-
-        if wks == None:
-            return None
-        
-        pis = wks.col_values(gform_info["sheet_data"]["pi_username_col"])
-        projects = wks.col_values(gform_info["sheet_data"]["project_name_col"])
-
-        all_responses = list(zip(pis, projects))
-        key = (self.request_obj.pi.username, self.request_obj.post_project.name)
-
-        if key not in all_responses:
-            return None
-        row_ind = all_responses.index(key) + 1
-
-        questions = wks.row_values(1)
-        response = wks.row_values(row_ind)
-
-        return zip(questions, response)
 
     def set_common_context_data(self, context):
         """Given a dictionary of context variables to include in the
